v3.7.0 for v3 andromo builder
Must use the new Andromo v3 builder to be able to target API 30  or higher so all the links were replaced with new one, for example the "andromo://.." links became "andromov3://.." and so forth.
Also, I've removed the RATE OUR APP button from the menus since because there is an notification popup option inside the Andromo now.